BUG-1365 : Reorganise code for readability
No functional changes
